To improve clarity, renamed JobOfferDetails model to JobOfferApplicationDetails and ProjectDetails model to ProjectApplicationDetails
