+++ application_app/models/job_offer_application_details.py
@@ -1,0 +1,6 @@
+from django.db import models
+
+
+class JobOfferApplicationDetails(models.Model):
+    cv = models.FileField()
+    motivation_letter = models.TextField()
