+++ application_app/models/project_application_details.py
@@ -1,6 +1,5 @@
 from django.db import models
 
 
-class JobOfferDetails(models.Model):
-    cv = models.FileField()
+class ProjectApplicationDetails(models.Model):
     motivation_letter = models.TextField()
