--- application_app/models/project_details.py
@@ -1,5 +1,0 @@
-from django.db import models
-
-
-class ProjectDetails(models.Model):
-    motivation_letter = models.TextField()
